HOCS-2502: change to force all 3 DCU permissions
At present the user selects one of the permissions and the team is
then added. This has the unintended issue that one team will not be
able to access the other 2 DCU areas. This change removes the case
type from the team creation and essentially hard codes the case
types of all 3 DCU areas.