HOCS-1983 Allow custom redirectURL on log out
Previously the logout URL was hardcoded to /, which actually meant the
root of the Keycloak server and not the URL of the Management UI
homepage.

This commit makes this editable, via an environment variable. I expect
this to be set in the kube config, usually to the $DOMAIN_NAME value,
although another option could be the URL of the non-management DECS.

We also rename "Logout" to "Log out", because it's a verb.