HOCS-1983 Remove redirect URL
The redirect URL on logout was defined incorrectly, and went to the root
of the Keycloak SSO server instead of the root of the MUI.

This commit removes the hardcoded redirect URL. A redirect URL will
still exist, but we'll pick it up from the values defined within
keycloak-proxy (aka keycloak-gatekeeper).

A downside to this approach is that if the redirect URL isn't defined in
keycloak-proxy it'll default to http://{value from host header} -- which
is mostly fine, but will involve using http:// briefly.

An alternative approach exists in PR #96 to specify explicit redirect
values, but that's adding complexity for minimal benefit. If we decide
to expose the MUI over multiple hostnames, that approach won't work.