HOCS-3744: double encode case reference
Add helper file to assist with double encoding case references. At
present when encoding a case reference of A/0...0/1 this turns to
`A%2F0...0%2F1`. When hocs-casework attempts to resolve the path, this
gets changed back to A/0...0/1 and fails to resolve to the correct path.
Double escaping the slash to %252F allows this to resolve correctly.